feat(#113926): Adding correct deserialization for JsonValue<Half>.
